M10

+++ Module 5/Extra oefeningen/Hoofdstuk 1.3 - Opgave 9.py
@@ -1,5 +1,5 @@
 def Ibol():
-    r = int(input("Geef de straal in cm:    "))
+    r = float(input("Geef de straal in cm:    "))
 
     I = round((4/3)*3.14159265359*(r**3))
     print(I)
